requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
gunicorn==21.2.0
python-binance==1.0.19
django-cors-headers==4.3.1
